server/requirements.txt
# FastAPI and Server dependencies
fastapi
uvicorn
python-multipart

# Gemini Book Extractor dependencies (merged from requirements_gemini.txt)
google-genai
Pillow
opencv-python
requests
redis
google-cloud-storage
pymongo
